intento traer menús por fecha pero hay problemas en el back. soluciono que se muestren los shifts nuevos en la lista al crearlos.

+++ src/app/modules/Dashboard/components/db-shifts-shift-list/db-shifts-shift-list.component.ts
@@ -1,7 +1,8 @@
-import { Component, inject } from '@angular/core';
+import { Component, Input, inject } from '@angular/core';
 import { Shift } from 'src/app/core/interfaces/shift.interface';
 import { ShiftsService } from 'src/app/core/services/shifts.service';
 import { Router } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-db-shifts-shift-list',
@@ -14,7 +15,9 @@ export class DbShiftsShiftListComponent {
 
   shiftsService = inject(ShiftsService)
 
-  router = inject(Router)
+  private router = inject(Router)
+
+  @Input() newShift!: Shift
 
   async ngOnInit() {
     try {
@@ -23,34 +26,67 @@ export class DbShiftsShiftListComponent {
       this.shifts = response
 
     } catch (error) {
-      console.log(error);      
+      console.log(error);
     }
   }
 
-  async erase(shift:any) {
-    try {
-      console.log("esto es shift.id");           
-      console.log(shift.id);      
+  async ngOnChanges() {
 
-      const response = await this.shiftsService.remove(shift.id)
-      console.log(response);
-      
-    } catch (error) {
-      console.log(error);      
-    }
+    const response = await this.shiftsService.create(this.newShift)
+    console.log(response);
+    this.shifts = await this.shiftsService.getAll()
   }
 
-  edit(shift:any) {
+  async erase(shift: any) {
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: "var(--secondary-color)",
+        cancelButton: "var(--third-color)"
+
+      },
+      buttonsStyling: false
+    });
+    swalWithBootstrapButtons.fire({
+      title: "¿Quieres eliminar este turno?",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Sí, eliminar",
+      cancelButtonText: "Cancelar",
+      reverseButtons: true,
+      background: "var(--bg-color)",
+      color: "var(--main-color)"
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        try {
+          console.log(shift.id);
+          const response = await this.shiftsService.remove(shift.id)
+          console.log(response);
+          this.shifts = await this.shiftsService.getAll()
+        } catch (error) {
+          console.log(error);
+        }
+        swalWithBootstrapButtons.fire({
+          title: "Reservation cancelled",
+          icon: "success",
+          confirmButtonColor: "var(--secondary-color)",
+          color: "var(--main-color)",
+          background: "(--bg-color)"
+        });
+      }
+    });
+  }
+
+  edit(shift: any) {
     const id = shift.id.toString()
-    this.router.navigate(["dashboard/shifts/edit",id])
+    this.router.navigate(["dashboard/shifts/edit", id])
 
     // try {
     //   const response = await this.shiftsService.update(shift)
     //   console.log(response);
-      
+
     // } catch(error) {
     //   console.log(error);
-      
+
     // }
   }
